Pull simulation into its own file
